Moved child reap from SIGCHLD handler to main loop
In this commit the child reap was moved from the SIGCHLD handler (sigchld_handler) to the main loop, right above the accept() call. The handler simply sets a flag notifying the main loop that there are children which might wait for attention and the main loop goes into the reap loop previously found in the handler.

This avoid possibly dangerous side-effects of calling process-related functions from within a signal handler.