Changed int to volatile sig_atomic_t in signal use
Changed the type of "sigchld_issued" from "int" to "volatile sig_atomic_t" which guarantees atomic writes from signal handlers. Also reordered the assignment to zero in the main loop to avoid sigchld_issued being set to 1 between the loop and the assignment to zero.

That wouldn't create much of an issue (the zombies would be killed in the next killing spree) and there is still the possibility of sigchld_issued being set to 1 by a SIGCHLD signal between its assignment to zero and the loop, but this time there won't be roaming zombies in the streets.